// src/services/model/getPostSlugs200.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * TrinityKitWP API
 * API para gerenciamento de posts, pedidos personalizados e formulários de contato no WordPress
 * OpenAPI spec version: 1.0.0
 */
import type { GetPostSlugs200DataItem } from './getPostSlugs200DataItem';

export type GetPostSlugs200 = {
  success?: boolean;
  data?: GetPostSlugs200DataItem[];
  /** Número total de posts existentes */
  total?: number;
  /** Número total de páginas disponíveis (quando usando paginação) */
  total_pages?: number;
  /** Número da página atual (quando usando paginação) */
  current_page?: number;
};