// src/services/model/getPostSlugs200DataItem.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * TrinityKitWP API
 * API para gerenciamento de posts, pedidos personalizados e formulários de contato no WordPress
 * OpenAPI spec version: 1.0.0
 */

export type GetPostSlugs200DataItem = {
  /** Slug do post */
  slug?: string;
  /** Título do post */
  title?: string;
  /** Conteúdo completo do post */
  content?: string;
  /** Resumo do post */
  excerpt?: string;
  /** Data de criação do post */
  created_at?: string;
  /** URL da imagem destacada do post */
  featured_image_url?: string;
};